# backend/src/main/java/com/srm/system/api/request/ParameterRequest.java
package com.srm.system.api.request;

public record ParameterRequest(String paramCode, String paramName, String paramType,
                               String defaultValue, String validationRule,
                               Boolean approvalRequired, String description) {}