# docs/04-接口与数据设计/openapi/internal-api.json
{
  "components": {
    "schemas": {
...
      "ApiResponseObject": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponsePageResultTaskView": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResultTaskView"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseString": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseVoid": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "NumberRuleRequest": {
        "properties": {
          "dateFormat": {
            "type": "string"
          },
          "objectType": {
            "type": "string"
          },
          "organizationDimension": {
            "type": "boolean"
          },
          "prefix": {
            "type": "string"
          },
          "resetCycle": {
            "type": "string"
          },
          "ruleCode": {
            "type": "string"
          },
          "ruleName": {
            "type": "string"
          },
          "serialLength": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "PageResultTaskView": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskView"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "pageSize": {
            "format": "int32",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          },
          "totalPages": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "ParameterRequest": {
        "properties": {
          "approvalRequired": {
            "type": "boolean"
          },
          "defaultValue": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "paramCode": {
            "type": "string"
          },
          "paramName": {
            "type": "string"
          },
          "paramType": {
            "type": "string"
          },
          "validationRule": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ParameterVersionRequest": {
        "properties": {
          "paramValue": {
            "type": "string"
          }
        },
        "required": [
          "paramValue"
        ],
        "type": "object"
      },
...
      "TaskView": {
        "properties": {
          "approvalInstanceId": {
            "format": "int64",
            "type": "integer"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "dueAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "priority": {
            "type": "string"
          },
          "slaStatus": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "sourceType": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "内部用户认证、RBAC、动态菜单、审批任务、系统治理及阶段1主数据接口。",
    "title": "SRM 内部管理端 API",
    "version": "v1"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/v1/master-data/organizations/{id}/disable": {
      "post": {
        "operationId": "disable_7",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用组织",
        "tags": [
          "organization-controller"
        ]
      }
    },
...
    "/api/v1/system/departments/{id}/disable": {
      "post": {
        "operationId": "disable_6",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用部门",
        "tags": [
          "department-controller"
        ]
      }
    },
...
    "/api/v1/system/dictionaries/{id}/disable": {
      "post": {
        "operationId": "disable_5",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Disable dictionary",
        "tags": [
          "dictionary-controller"
        ]
      }
    },
...
    "/api/v1/system/document-templates/{id}/disable": {
      "post": {
        "operationId": "disable_4",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Disable document template",
        "tags": [
          "document-template-controller"
        ]
      }
    },
...
    "/api/v1/system/number-rules": {
      "get": {
        "operationId": "list_4",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List number rule",
        "tags": [
          "number-rule-controller"
        ]
      },
      "post": {
        "operationId": "create_5",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NumberRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Create number rule",
        "tags": [
          "number-rule-controller"
        ]
      }
    },
    "/api/v1/system/number-rules/generate": {
      "post": {
        "operationId": "generate",
        "parameters": [
          {
            "in": "query",
            "name": "ruleCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "orgId",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Generate number rule",
        "tags": [
          "number-rule-controller"
        ]
      }
    },
    "/api/v1/system/number-rules/{id}": {
      "put": {
        "operationId": "update_4",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NumberRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Update number rule",
        "tags": [
          "number-rule-controller"
        ]
      }
    },
    "/api/v1/system/number-rules/{id}/disable": {
      "post": {
        "operationId": "disable_3",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Disable number rule",
        "tags": [
          "number-rule-controller"
        ]
      }
    },
    "/api/v1/system/number-rules/{id}/enable": {
      "post": {
        "operationId": "enable_3",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Enable number rule",
        "tags": [
          "number-rule-controller"
        ]
      }
    },
...
    "/api/v1/system/parameters": {
      "get": {
        "operationId": "list_3",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List parameter",
        "tags": [
          "parameter-controller"
        ]
      },
      "post": {
        "operationId": "create_4",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParameterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Create parameter",
        "tags": [
          "parameter-controller"
        ]
      }
    },
    "/api/v1/system/parameters/{id}": {
      "get": {
        "operationId": "get_7",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get parameter",
        "tags": [
          "parameter-controller"
        ]
      }
    },
    "/api/v1/system/parameters/{pid}/versions": {
      "get": {
        "operationId": "versions",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List versions parameter",
        "tags": [
          "parameter-controller"
        ]
      },
      "post": {
        "operationId": "createVersion",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParameterVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Create parameter",
        "tags": [
          "parameter-controller"
        ]
      }
    },
    "/api/v1/system/parameters/{pid}/versions/{vid}/submit": {
      "post": {
        "operationId": "submit",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "vid",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Submit parameter",
        "tags": [
          "parameter-controller"
        ]
      }
    },
...
    "/api/v1/system/positions/{id}/disable": {
      "post": {
        "operationId": "disable_2",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用岗位",
        "tags": [
          "position-controller"
        ]
      }
    },
...
    "/api/v1/system/roles/{id}/disable": {
      "post": {
        "operationId": "disable_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用角色",
        "tags": [
          "role-controller"
        ]
      }
    },
...
    "/api/v1/system/users/{id}/disable": {
      "post": {
        "operationId": "disable",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用用户",
        "tags": [
          "user-controller"
        ]
      }
    },
...
    "/api/v1/tasks/my": {
      "get": {
        "operationId": "myTasks",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 10,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResultTaskView"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "我的任务分页",
        "tags": [
          "task-controller"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "通过当前站点的同源网关访问",
      "url": "/"
    }
  ]
}